feat: Move partial stuff into own files
