+++ src/chat.py
@@ -1,0 +1,9 @@
+from openai.resources.chat import Chat
+
+from completions import AiRouterCompletions
+
+
+class AiRouterChat(Chat):
+    @property
+    def completions(self):
+        return AiRouterCompletions(self._client)
